// packages/contracts/src/enums/analytics-metric.enum.ts
export enum AnalyticsMetric {
  VIEWS = 'views',
  VIDEO_VIEWS = 'videoViews',
  IMPRESSIONS = 'impressions',
  REACH = 'reach',
  LIKES = 'likes',
  COMMENTS = 'comments',
  SHARES = 'shares',
  SAVES = 'saves',
  CLICKS = 'clicks',
  WATCH_TIME = 'watchTime',
  FOLLOWERS = 'followers',
  SUBSCRIBERS = 'subscribers',
  ENGAGEMENT_RATE = 'engagementRate',
  POSTS = 'posts',
  ENGAGEMENT = 'engagement',
}